Add once_per_grid and once_per_block helper functions
`once_per_grid(acc)` returns true for a single thread within the kernel
execution grid. Usually the condition is true for block 0 and thread 0,
but these indices should not be relied upon.

`once_per_block(acc)` returns true for a single thread within the block.
Usually the condition is true for thread 0, but this index should not be
relied upon.